+++ web/src/core/BeamUtils.ts
@@ -15,10 +15,12 @@ import type { Beam } from '@/core/Beam';
 import type { BeamServer } from '@/core/BeamServer';
 import {
   AccountService,
+  AnalyticsService,
   AnnouncementsService,
   AuthService,
   ContentService,
   LeaderboardsService,
+  MailService,
   MessageRailService,
   StatsService,
 } from '@/services';
@@ -113,10 +115,12 @@ export function clientServices(beam: Beam) {
   const apiServiceProps = { beam, getPlayer: () => beam.player };
   const allApiServices = [
     new AccountService(apiServiceProps),
+    new AnalyticsService(apiServiceProps),
     new AnnouncementsService(apiServiceProps),
     new AuthService(apiServiceProps),
     new ContentService(apiServiceProps),
     new LeaderboardsService(apiServiceProps),
+    new MailService(apiServiceProps),
     new MessageRailService(apiServiceProps),
     new StatsService(apiServiceProps),
   ];
@@ -137,10 +141,12 @@ export function serverServices(beamServer: BeamServer) {
   const apiServiceProps = { beam: beamServer };
   const allApiServices = [
     new AccountService(apiServiceProps),
+    new AnalyticsService(apiServiceProps),
     new AnnouncementsService(apiServiceProps),
     new AuthService(apiServiceProps),
     new ContentService(apiServiceProps),
     new LeaderboardsService(apiServiceProps),
+    new MailService(apiServiceProps),
     new MessageRailService(apiServiceProps),
     new StatsService(apiServiceProps),
   ];